libodroid-gps: add to scan USB GPS dongle and find a device
When USB GPS dongle is connected with other USB dongles, its device name may
not be same as the one which is subscribed in the system property
ro.kernel.android.gps' such as '/dev/ttyACM0'. Then GPS device cannot be
opened due to device name is not matched.

So this patch is to scan the USB GPS dongle with idVendor and idProduct listed
in the file 'odroid-usbgps.xml' on USB buses, and find decent device name.
Please note that this device name is not regular one such as '/dev/ttyACM0',
instead it's like '/dev/bus/usb/002/005' which is based on USB bus number and
device number as connected.

You can add new USB GPS dongle simply with its idVendor and idProduct as well
as its baudrate.

    <devices>
        <usbdev vid="1546" pid="01a6" baudrate="9600" />
        <usbdev vid="10c4" pid="ea60" baudrate="9600" />
    </devices>

Change-Id: If6c38fe78b041fc5f7ec618a05f1d9ae54b23418
Signed-off-by: Dongjin Kim <tobetter@gmail.com>